+++ ta4j-core/src/main/java/org/ta4j/core/analysis/CashFlow.java
@@ -62,8 +62,17 @@ public class CashFlow implements Indicator<Decimal> {
      * @param tradingRecord the trading record
      */
     public CashFlow(TimeSeries timeSeries, TradingRecord tradingRecord) {
+        this(timeSeries, tradingRecord, timeSeries.getBeginIndex(), timeSeries.getEndIndex());
+    }
+
+    /**
+     * Constructor.
+     * @param timeSeries the time series
+     * @param tradingRecord the trading record
+     */
+    public CashFlow(TimeSeries timeSeries, TradingRecord tradingRecord, int beginIndex, int endIndex) {
         this.timeSeries = timeSeries;
-        calculate(tradingRecord);
+        calculate(tradingRecord, beginIndex, endIndex);
         fillToTheEnd();
     }
 
@@ -115,8 +124,11 @@ public class CashFlow implements Indicator<Decimal> {
      * Calculates the cash flow for a trading record.
      * @param tradingRecord the trading record
      */
-    private void calculate(TradingRecord tradingRecord) {
+    private void calculate(TradingRecord tradingRecord, int beginIndex, int endIndex) {
         for (Trade trade : tradingRecord.getTrades()) {
+            if (trade.getEntry().getIndex() < beginIndex || trade.getExit().getIndex() > endIndex) {
+                continue;
+            }
             // For each trade...
             calculate(trade);
         }