+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/BuyAndHoldCriterion.java
@@ -35,7 +35,12 @@ public class BuyAndHoldCriterion extends AbstractAnalysisCriterion {
 
     @Override
     public double calculate(TimeSeries series, TradingRecord tradingRecord) {
-        return series.getBar(series.getEndIndex()).getClosePrice().dividedBy(series.getBar(series.getBeginIndex()).getClosePrice()).doubleValue();
+        return calculate(series, tradingRecord, series.getBeginIndex(), series.getEndIndex());
+    }
+
+    @Override
+    public double calculate(TimeSeries series, TradingRecord tradingRecord, int beginIndex, int endIndex) {
+        return series.getBar(endIndex).getClosePrice().dividedBy(series.getBar(beginIndex).getClosePrice()).doubleValue();
     }
 
     @Override